requirements.txt
aiodns==1.1.1
aiohttp==3.1.3
fire==0.1.3
pyyaml==3.12
